+++ js/benchmarks/evals-benchmarks/evals/faithfulness.eval.ts
@@ -26,7 +26,7 @@ import { openai } from "@ai-sdk/openai";
 import * as px from "@arizeai/phoenix-client/vitest";
 import { createFaithfulnessEvaluator } from "@arizeai/phoenix-evals";
 
-import { benchmarkSuite, labelAccuracy } from "../src/meta-evaluators";
+import { benchmarkSuite, labelAccuracy } from "../src/evaluators";
 
 const faithfulnessEvaluator = createFaithfulnessEvaluator({
   model: openai("gpt-4o-mini"),
