+++ js/benchmarks/evals-benchmarks/evals/refusal.eval.ts
@@ -18,7 +18,7 @@ import { openai } from "@ai-sdk/openai";
 import * as px from "@arizeai/phoenix-client/vitest";
 import { createRefusalEvaluator } from "@arizeai/phoenix-evals";
 
-import { benchmarkSuite, labelAccuracy } from "../src/meta-evaluators";
+import { benchmarkSuite, labelAccuracy } from "../src/evaluators";
 
 const refusalEvaluator = createRefusalEvaluator({
   model: openai("gpt-4o-mini"),
