+++ js/benchmarks/evals-benchmarks/evals/tool_response_handling.eval.ts
@@ -20,7 +20,7 @@ import { openai } from "@ai-sdk/openai";
 import * as px from "@arizeai/phoenix-client/vitest";
 import { createToolResponseHandlingEvaluator } from "@arizeai/phoenix-evals";
 
-import { benchmarkSuite, labelAccuracy } from "../src/meta-evaluators";
+import { benchmarkSuite, labelAccuracy } from "../src/evaluators";
 
 const toolResponseHandlingEvaluator = createToolResponseHandlingEvaluator({
   model: openai("gpt-4o-mini"),
